pyproject.toml
[tool.poetry]
name = "online-training"
version = "0.1.0"
description = ""
authors = ["KudryashovR <romkud55@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
